+++ src/mainPage/MathProblemsSection.tsx
@@ -1,0 +1,15 @@
+import { Accordion } from "react-bootstrap";
+import TopicSection from "./TopicSection";
+import topics from "./topics-names-list.json";
+
+export default function MathProblemsSection() {
+  return (
+    <div>
+      <Accordion>
+        {topics.map((topic) => (
+          <TopicSection topic={topic} />
+        ))}
+      </Accordion>
+    </div>
+  );
+}